decompose namespace & definition & definition_block & statement_block
