+++ include/ast/statements/conditional_branch.hpp
@@ -2,6 +2,7 @@
 
 #include "ast/expressions/expr.hpp"
 #include "ast/statements/statement.hpp"
+#include "ast/statements/statement_block.hpp"
 
 class conditional_branch : public statement {
 public:
