+++ include/parser/namespace/parse_namespace.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
 #include "parser/parse_defination.hpp"
+#include "ast/namespace.hpp"
 
 TYPE_namespace* parse_namespace(token_list& Token_list);
